Update command keept together with other commands
