+++ model/sharedStimulus/factory/CommandFactory.php
@@ -22,9 +22,12 @@ declare(strict_types=1);
 
 namespace oat\taoMediaManager\model\sharedStimulus\factory;
 
+use _HumbugBox7f2450e54e51\Psr\Log\InvalidArgumentException;
 use oat\oatbox\service\ConfigurableService;
+use oat\oatbox\user\User;
 use oat\taoMediaManager\model\MediaService;
 use oat\taoMediaManager\model\sharedStimulus\CreateCommand;
+use oat\taoMediaManager\model\sharedStimulus\UpdateCommand;
 use Psr\Http\Message\ServerRequestInterface;
 
 class CommandFactory extends ConfigurableService
@@ -37,6 +40,21 @@ class CommandFactory extends ConfigurableService
             $parsedBody['classId'] ?? $parsedBody['classUri'] ?? MediaService::ROOT_CLASS_URI,
             $parsedBody['name'] ?? null,
             $parsedBody['languageId'] ?? $parsedBody['languageUri'] ?? null
+        );
+    }
+
+    public function patchStimulusByRequest(ServerRequestInterface $request, User $user): UpdateCommand
+    {
+        $parsedBody = json_decode((string)$request->getBody(), true);
+
+        if (JSON_ERROR_NONE !== json_last_error()) {
+            throw new InvalidArgumentException('Incorrect request format');
+        }
+
+        return new UpdateCommand(
+            $parsedBody['id'],
+            $parsedBody['body'],
+            $user->getIdentifier()
         );
     }
 }
